Bug 508725 - Part 2: Record on nsCSSStyleSheets whether they are for a <style scoped>. r=dbaron
UltraBlame original commit: b3c6cd8f1c5600decae9eb340ef626bec6acbdf8

+++ layout/style/nsRuleData.h
@@ -25,7 +25,7 @@ struct nsRuleData
   const uint32_t mSIDs;
   bool mCanStoreInRuleTree;
   bool mIsImportantRule;
-  uint8_t mLevel; 
+  uint16_t mLevel; 
   nsPresContext* const mPresContext;
   nsStyleContext* const mStyleContext;
   const nsPostResolveFunc mPostResolveCallback;
